fix: add validation for admin and user owner
Add validations for admin and user owner

+++ groups/src/main/java/edu/pjwstk/groups/usecase/deletegroupinvitation/DeleteGroupInvitationByIdImpl.java
@@ -1,6 +1,10 @@
 package edu.pjwstk.groups.usecase.deletegroupinvitation;
 
+import edu.pjwstk.common.authApi.AuthApi;
+import edu.pjwstk.common.authApi.dto.CurrentUserDto;
+import edu.pjwstk.groups.entity.GroupInvitation;
 import edu.pjwstk.groups.exception.GroupInvitationNotFoundException;
+import edu.pjwstk.groups.exception.UserNotGroupAdministratorAccessDeniedException;
 import edu.pjwstk.groups.repository.GroupInvitationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -8,20 +12,30 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.UUID;
 
 @Service
-public class DeleteGroupInvitationByIdImpl implements DeleteGroupInvitationById{
+public class DeleteGroupInvitationByIdImpl implements DeleteGroupInvitationById {
 
     private final GroupInvitationRepository groupInvitationRepository;
+    private final AuthApi authApi;
 
-    public DeleteGroupInvitationByIdImpl(GroupInvitationRepository groupInvitationRepository) {
+    public DeleteGroupInvitationByIdImpl(GroupInvitationRepository groupInvitationRepository, AuthApi authApi) {
         this.groupInvitationRepository = groupInvitationRepository;
+        this.authApi = authApi;
     }
 
     @Override
     @Transactional
     public void execute(UUID groupInvitationId) {
-        groupInvitationRepository.findById(groupInvitationId)
-                        .orElseThrow(() -> new GroupInvitationNotFoundException("Group invitation with id:"
-                                + groupInvitationId + " not found!"));
+        GroupInvitation groupInvitation = groupInvitationRepository.findById(groupInvitationId)
+                .orElseThrow(() -> new GroupInvitationNotFoundException("Group invitation with id:"
+                        + groupInvitationId + " not found!"));
+
+        CurrentUserDto currentUserDto = authApi.getCurrentUser()
+                .orElseThrow();
+
+        if (currentUserDto.userId() != groupInvitation.getGroupInvited().getAdminId()) {
+            throw new UserNotGroupAdministratorAccessDeniedException("Only group administrators can delete group invitations!");
+        }
+
         groupInvitationRepository.deleteById(groupInvitationId);
     }
 }
